postgres_to_es/requirements.txt
psycopg2-binary==2.9.1
python-dotenv==0.19.0
requests==2.26.0
redis==3.5.3
pydantic==1.8.2